requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
requests==2.32.3
geopy==2.4.1
python-dotenv==1.0.1
pydantic>=1.10.5
email-validator>=1.3.0
gunicorn==20.1.0
httpx>=0.24.1
pytest>=7.3.1
pytest-asyncio>=0.20.1
black>=23.1.0
mypy>=0.991